Remove Dependency on Fman Build System (FBS) for Packaging GUI (#181)
fbs (https://github.com/mherrmann/fbs) seems to not be supported very much anymore, and I was using a paid version of it that was frozen to a particular version of it, so I couldn't pull in new changes nor was it great having my project depend on a binary that I couldn't publicly share.

Ripping it out entailed a few things:

Now using pyinstaller directly (No longer having to be pinned to an old version of it either)
Creating the installers for Windows, Linux, and Mac manually
Replacing a lot of the helper config/settings with a new app_settings.py file
And additional refactorings I did in the process:

GUI no longer uses bundled CLI executable -- now just calls CLI module directly
Lots of workflow optimizations, including a new workflow build-and-verify.yml that is a parent workflow for the other three.
No more annoying _version.py file or bump_versions.sh script or freeze-prep.sh script.
Nothing about the application itself should have changed from a user perspective (except the logging output may look a little different in some areas), but I need to watch for bugs and hiccups for a little while on my own before tagging a new version.

+++ gui/src/main/python/progress_dialog.py
@@ -14,4 +14,10 @@ class ProgressDialog(QDialog, Ui_ProgressDialog):
     @Slot()
     def abort(self):
         self.parentWidget().message("Abort button clicked!")
-        self.parentWidget().p.kill()
+        if hasattr(self.parentWidget(), "p") and self.parentWidget().p is not None:
+            self.parentWidget().p.kill()
+        elif hasattr(self.parentWidget(), "_internal_worker"):
+            self.parentWidget()._internal_worker.cancel()
+            self.parentWidget().message("Cancelling operation...")
+        else:
+            self.parentWidget().message("Abort not supported for this operation.")
